Tighten TypeScript types and casts
Replace unsafe non-null assertions with explicit type assertions and add missing type imports across transaction modules and tests. Changes include casting PSBT inputs/outputs and signatures (as PsbtInput, as PsbtOutputExtended, as Uint8Array, as string[], as UTXO, etc.), importing PsbtInput, adjusting witness/signature handling, and annotating FastBigIntMap iterators and callbacks. Also update test import path to .js and correct a toHex call with an explicit Uint8Array cast. These are typing-only fixes to satisfy stricter TypeScript settings and do not change runtime logic.

+++ test/old/FastBigIntMap.ts
@@ -91,7 +91,7 @@ export class FastBigIntMap {
      */
     *entries(): IterableIterator<[bigint, bigint]> {
         for (const key of this.keyOrder) {
-            yield [key, this.items[key.toString()]];
+            yield [key, this.items[key.toString()] as bigint];
         }
     }
 
@@ -107,7 +107,7 @@ export class FastBigIntMap {
      */
     *values(): IterableIterator<bigint> {
         for (const key of this.keyOrder) {
-            yield this.items[key.toString()];
+            yield this.items[key.toString()] as bigint;
         }
     }
 
@@ -119,7 +119,7 @@ export class FastBigIntMap {
         thisArg?: unknown,
     ): void {
         for (const key of this.keyOrder) {
-            callback.call(thisArg, this.items[key.toString()], key, this);
+            callback.call(thisArg, this.items[key.toString()] as bigint, key, this);
         }
     }
